Update pylint configuration for v33
I don't know if all of these are new since pylint 1.8 (Ubuntu 18.04)
—some of them certainly are— but in any case I have seen them in real
test-packs and I disagree with them because they're either too coarse or
too opinionated, or (in the case of `import-outside-toplevel`) interfere
with patterns we use (for page object navigation methods that return
another page object we often need cyclic imports, so importing in the
method that's going to return it is the only way).

stbt_rig.py is a file maintained by Stb-tester.com, that needs to keep
compatibility with Python 2 & 3 simultaneously, so it fails some of
pylint's checks which expect Python 3 only. We don't need to lint it
here because it's maintained by Stb-tester.com and linted in its
upstream CI: https://github.com/stb-tester/stbt-rig